added code from my space Invaders project to this one.
most of the code was just copied and pasted to this project.
all of the code files were added to the codeblocks project file.
none of the code was integrated into the application.
meaning the space invaders project is actually doing anything.

second added a simple script and post build step to
run the natural documentation generator.

none of the code has actually been documented it yet.