added and moved methods for the tetris gameManager to public
createBlock and CreateRandomBlock were added to the tetris
gameManager class. the have already been documented
for natural docs. also exposed an blockColors enum in the tetrisGameManager.
I still need to finish documenting the blockColors enum in the tetrisGameManager.
and use a shape enum instead of color Enum for parts of tetrisGameManger.