moved shoot test to player class from playermanager
the reason for this is so when playermanager
and enemymanager get merged into one class
that player specfic is already in player.
also moved the timer member in playermanager  to player class.
O

note:
it may be a good idea to move the adding player bullet code in playermanager
to player class so that the playermanager doesnt have the responsibility.